+++ src/main/java/com/atlassian/db/replica/spi/DataSource.java
@@ -1,0 +1,7 @@
+package com.atlassian.db.replica.spi;
+
+import java.sql.Connection;
+
+public interface DataSource {
+    Connection getConnection();
+}