+++ src/test/java/com/atlassian/db/replica/internal/LsnReplicaConsistencyTest.java
@@ -22,7 +22,7 @@ public class LsnReplicaConsistencyTest {
         final ReplicaConsistency consistency = new LsnReplicaConsistency();
         consistency.write(getConnection("16/3002D50"));
 
-        final boolean isConsistent = consistency.isConsistent(getConnectionSupplier("16/3002D50"));
+        final boolean isConsistent = consistency.isConsistent(() -> getConnectionSupplier("16/3002D50"));
 
         assertThat(isConsistent).isTrue();
     }
@@ -35,7 +35,7 @@ public class LsnReplicaConsistencyTest {
 
         consistency.write(main);
 
-        final boolean isConsistent = consistency.isConsistent(getConnectionSupplier("16/3002D50"));
+        final boolean isConsistent = consistency.isConsistent(() -> getConnectionSupplier("16/3002D50"));
 
         assertThat(isConsistent).isFalse();
     }
@@ -49,7 +49,7 @@ public class LsnReplicaConsistencyTest {
         consistency.write(main);
         consistency.write(getConnection("16/3002D50"));
 
-        final boolean isConsistent = consistency.isConsistent(getConnectionSupplier("16/3002D50"));
+        final boolean isConsistent = consistency.isConsistent(() -> getConnectionSupplier("16/3002D50"));
 
         assertThat(isConsistent).isTrue();
     }
@@ -57,9 +57,9 @@ public class LsnReplicaConsistencyTest {
     @Test
     public void shouldNotBeConsistentBeforeTheFirstWrite() {
         final ReplicaConsistency consistency = new LsnReplicaConsistency();
-        final Supplier<Connection> replica = mock(ConnectionSupplier.class);
+        final ConnectionSupplier replica = mock(ConnectionSupplier.class);
 
-        final boolean isConsistent = consistency.isConsistent(replica);
+        final boolean isConsistent = consistency.isConsistent(() -> replica);
 
         assertThat(isConsistent).isFalse();
     }
@@ -72,13 +72,17 @@ public class LsnReplicaConsistencyTest {
         final Connection replica = mock(Connection.class);
         when(replica.prepareStatement(anyString())).thenThrow(new SQLException("Replica connection fails"));
 
-        final boolean isConsistent = consistency.isConsistent(new ConnectionSupplier(replica));
+        final boolean isConsistent = consistency.isConsistent(() -> new ConnectionSupplier(replica));
 
         assertThat(isConsistent).isFalse();
     }
 
-    private Supplier<Connection> getConnectionSupplier(String lsn) throws SQLException {
-        return new ConnectionSupplier(getConnection(lsn));
+    private ConnectionSupplier getConnectionSupplier(String lsn) {
+        try {
+            return new ConnectionSupplier(getConnection(lsn));
+        } catch (SQLException e) {
+            throw new RuntimeException(e);
+        }
     }
 
     private Connection getConnection(String lsn) throws SQLException {